Se realizan pruebas de funcionalidad de filtros en archivo home.js. Se elimina el archivo Example.js
