--- src/views/Example.js
@@ -1,7 +1,0 @@
-
-
-
-
-
-
-
